Moved the lua tests to the redis tests class.
